Update Core.py
Began moving some of the repeating request portions out to class properties. This has made it significantly easier to read and easier to write. worked on a few more update functions. considering whether City().update_short() should just be update() considering it is not used or whether to stick to the convention elsewhere. Also set up a few todo comments for some things that need to be done.